+++ lib/support/live_select_web/live/live_component_form.ex
@@ -50,8 +50,8 @@ defmodule LiveSelectWeb.LiveComponentForm do
           (form created in .form component, captured with :let)
         </a>
       </div>
-      <.form :let={f} for={%{}} as={:my_form_old_style} phx-submit="submit" phx-target={@myself}>
-        <.live_select form={f} field={:city_search} mode={:tags} />
+      <.form :let={f} for={%{}} as={:my_form_old_style} phx-submit="submit">
+        <.live_select form={f} field={:city_search} mode={:tags} phx-target={@myself} />
         <%= submit("Submit", class: "btn btn-primary") %>
       </.form>
     </div>